# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GreenLink & Sunrise Agrovet Retail Operations OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
          description: Filter by date (YYYY-MM-DD), defaults to today
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Record a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Sale recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{saleId}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale with items
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithItems"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Remove sale (restores stock)
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteSaleInput"
      responses:
        "204":
          description: Sale removed

  # ─── DEBTS ───────────────────────────────────────────────
...
  /reports/dashboard:
    get:
      operationId: getDashboard
      tags: [reports]
      summary: Dashboard summary for today or a date
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
          description: YYYY-MM-DD, defaults to today
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /reports/range:
    get:
      operationId: getReportRange
      tags: [reports]
      summary: Sales report for a date range
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Range report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RangeReport"
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Top products by revenue or profit
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /reports/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [reports]
      summary: Revenue and profit breakdown by category
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
  /reports/hourly:
    get:
      operationId: getHourlySales
      tags: [reports]
      summary: Sales by hour for a given date
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Hourly breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourlySales"
  /reports/debts-summary:
    get:
      operationId: getDebtsSummary
      tags: [reports]
      summary: Debt summary totals
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Debts summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtsSummary"

  # ─── OCR ─────────────────────────────────────────────────
...
components:
  schemas:
...
    SaleItem:
      type: object
      properties:
        id:
          type: string
        saleId:
          type: string
        productId:
          type: ["string", "null"]
        productName:
          type: string
        qty:
          type: integer
        unitPrice:
          type: number
        unitCost:
          type: ["number", "null"]
        unitProfit:
          type: ["number", "null"]
        totalPrice:
          type: number
        totalProfit:
          type: ["number", "null"]
      required: [id, saleId, productName, qty, unitPrice, totalPrice]
    Sale:
      type: object
      properties:
        id:
          type: string
        shopId:
          type: string
        totalAmount:
          type: number
        totalCost:
          type: ["number", "null"]
        totalProfit:
          type: ["number", "null"]
        discount:
          type: number
        saleType:
          type: string
          enum: [cash, debt]
        servedBy:
          type: ["string", "null"]
        syncStatus:
          type: string
        createdAt:
          type: string
      required: [id, shopId, totalAmount, discount, saleType, syncStatus, createdAt]
    SaleWithItems:
      type: object
      properties:
        id:
          type: string
        shopId:
          type: string
        totalAmount:
          type: number
        totalCost:
          type: ["number", "null"]
        totalProfit:
          type: ["number", "null"]
        discount:
          type: number
        saleType:
          type: string
        servedBy:
          type: ["string", "null"]
        syncStatus:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
      required: [id, shopId, totalAmount, discount, saleType, syncStatus, createdAt, items]
    SaleItemInput:
      type: object
      properties:
        productId:
          type: string
        qty:
          type: integer
        unitPrice:
          type: number
        discount:
          type: ["number", "null"]
      required: [productId, qty, unitPrice]
    SaleInput:
      type: object
      properties:
        shopId:
          type: string
        saleType:
          type: string
          enum: [cash, debt]
        servedBy:
          type: ["string", "null"]
        discount:
          type: ["number", "null"]
        discountOverrideBy:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
        debtCustomerName:
          type: ["string", "null"]
        debtCustomerPhone:
          type: ["string", "null"]
      required: [shopId, saleType, items]
    DeleteSaleInput:
      type: object
      properties:
        reason:
          type: ["string", "null"]
        performedBy:
          type: ["string", "null"]

    # ─── DEBTS ────────────────────────────────────────────
...
    DashboardSummary:
      type: object
      properties:
        date:
          type: string
        shopId:
          type: ["string", "null"]
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalCost:
          type: number
        salesCount:
          type: integer
        marginPct:
          type: number
        cashSales:
          type: number
        debtSales:
          type: number
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/TopProduct"
        lowStockCount:
          type: integer
        outOfStockCount:
          type: integer
        pendingDebtsTotal:
          type: number
        unreadNotificationsCount:
          type: integer
      required: [date, totalRevenue, totalProfit, totalCost, salesCount, marginPct, cashSales, debtSales, topProducts, lowStockCount, outOfStockCount, pendingDebtsTotal, unreadNotificationsCount]
    RangeReport:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        shopId:
          type: ["string", "null"]
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalCost:
          type: number
        salesCount:
          type: integer
        marginPct:
          type: number
        dailyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DailySummary"
      required: [from, to, totalRevenue, totalProfit, totalCost, salesCount, marginPct, dailyBreakdown]
    DailySummary:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        profit:
          type: number
        salesCount:
          type: integer
      required: [date, revenue, profit, salesCount]
    TopProduct:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        totalQtySold:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: ["number", "null"]
      required: [productId, productName, totalQtySold, totalRevenue]
    CategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        totalRevenue:
          type: number
        totalProfit:
          type: number
        salesCount:
          type: integer
      required: [category, totalRevenue, totalProfit, salesCount]
    HourlySales:
      type: object
      properties:
        hour:
          type: integer
        salesCount:
          type: integer
        revenue:
          type: number
      required: [hour, salesCount, revenue]
    DebtsSummary:
      type: object
      properties:
        totalOutstanding:
          type: number
        totalUnpaid:
          type: number
        totalPartial:
          type: number
        unpaidCount:
          type: integer
        partialCount:
          type: integer
      required: [totalOutstanding, totalUnpaid, totalPartial, unpaidCount, partialCount]

    # ─── OCR ──────────────────────────────────────────────
...